src/normal_nds/requirements.txt
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html

torch
torchvision
torchaudio
ninja
imageio==2.9.0
imageio-ffmpeg==0.4.4
opencv-python==4.5.2.54
numpy==1.22.3
matplotlib==3.4.3
meshzoo==0.9.11
pillow==8.1.0 
scikit-image==0.18.1
tensorboard==2.2.0
tqdm==4.56.2
trimesh==3.9.20
